refactor: fragment 별 디렉토리 구조로 변경
