+++ app/src/main/java/com/survivalcoding/stopwatch/presentation/bedtime/BedtimeFragment.kt
@@ -1,4 +1,4 @@
-package com.survivalcoding.stopwatch.presentation.fragment
+package com.survivalcoding.stopwatch.presentation.bedtime
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -15,11 +15,11 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [TimerFragment.newInstance] factory method to
+ * Use the [BedtimeFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
 @AndroidEntryPoint
-class TimerFragment : Fragment() {
+class BedtimeFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -37,7 +37,7 @@ class TimerFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_timer, container, false)
+        return inflater.inflate(R.layout.fragment_bedtime, container, false)
     }
 
     companion object {
@@ -47,12 +47,12 @@ class TimerFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment timerFragment.
+         * @return A new instance of fragment bedtimeFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            TimerFragment().apply {
+            BedtimeFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
