--- app/src/main/java/com/survivalcoding/stopwatch/presentation/state/ProgressBarState.kt
@@ -1,5 +1,0 @@
-package com.survivalcoding.stopwatch.presentation.state
-
-data class ProgressBarState(
-    val percent: Int = 0
-)
