+++ app/src/main/java/com/survivalcoding/stopwatch/presentation/stopwatch/state/ProgressBarState.kt
@@ -1,0 +1,5 @@
+package com.survivalcoding.stopwatch.presentation.stopwatch.state
+
+data class ProgressBarState(
+    val percent: Int = 0
+)
