+++ app/src/main/res/layout/fragment_alarm.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.fragment.AlarmFragment">
+    tools:context=".presentation.alarm.AlarmFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
