+++ app/src/main/res/layout/fragment_stop_watch.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background_main"
-    tools:context=".presentation.fragment.StopWatchFragment">
+    tools:context=".presentation.stopwatch.StopWatchFragment">
 
     <androidx.constraintlayout.motion.widget.MotionLayout
         android:id="@+id/progressive_timer_button_wrapper"
